phase8_streamlit/requirements.txt
streamlit==1.34.0
pandas==2.2.2
pyarrow==15.0.2
python-dotenv==1.0.1
